Bidding process GUI

+++ src/main/java/ba/atlantbh/auctionapp/responses/SimpleSubcategoryResponse.java
@@ -1,5 +1,6 @@
 package ba.atlantbh.auctionapp.responses;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface SimpleSubcategoryResponse {
@@ -7,5 +8,5 @@ public interface SimpleSubcategoryResponse {
     String getName();
     String getCategoryName();
     String getUrl();
-    Integer getStartPrice();
+    BigDecimal getStartPrice();
 }
